Put specs in a single file. We want to be sure mock tests run last, and relying on filesystem ordering seems fragile to me.
