+++ config/ci.yml.sample
@@ -1,3 +1,4 @@
+# NOTE: These sample values are used by the mock tests.
 username: my-name@my-place.com
 password: 'my-password'
 client_id:     32-hex-digit-id
